+++ internal/tools/crd/crd.go
@@ -1,4 +1,4 @@
-package tools
+package crd
 
 import (
 	"context"
