+++ internal/tools/rbactools/serviceaccount.go
@@ -1,4 +1,4 @@
-package tools
+package rbactools
 
 import (
 	"context"
