--- internal/tools/tools_test.go
@@ -1,26 +1,0 @@
-//go:build integration
-// +build integration
-
-package tools_test
-
-import (
-	"os"
-	"path"
-
-	"k8s.io/client-go/tools/clientcmd"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-func setupKubeClient() (client.Client, error) {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return nil, err
-	}
-
-	cfg, err := clientcmd.BuildConfigFromFlags("", path.Join(home, ".kube/config"))
-	if err != nil {
-		return nil, err
-	}
-
-	return client.New(cfg, client.Options{})
-}
